# docs/openapi.yaml
openapi: 3.0.3
...
info:
  title: Mova – Gym Tracker API
  description: |
    REST API for the **Mova** gym-tracking web application.

    ## Authentication
    All endpoints (except registration) require **HTTP Basic Auth**.
    Include your credentials via the `Authorization: Basic <base64(username:password)>` header.

    ## Roles
    | Role    | Description |
    |---------|-------------|
    | `admin` | Manages exercises, templates, and exercise requests |
    | `user`  | Logs workouts, tracks metrics, manages personal plans |

    ## Business Rules
    1. **Exercise Authority Integrity** — Only admins can create/update/delete exercises. Deletion is a soft-delete (`is_active = false`) to preserve historical data.
    2. **Anti-Spam Security** — A user can have at most **5 pending** exercise requests at any time.
    3. **Data Validation** — A workout session cannot be saved unless it contains at least one workout set.
  version: 1.0.0
  contact:
    name: Team Mova
servers:
  - url: http://localhost:8000
    description: Local Development
  - url: https://api.mova.app
    description: Production

# =============================================================================
#  Global Security — Basic Auth on every endpoint by default
# =============================================================================
security:
  - basicAuth: []
...
paths:
...
  /workout-plans:
    get:
      tags: [Workout Plans]
      summary: List workout plans
      description: |
        Returns all workout plans visible to the current user.
        - Public templates (`is_template = true`) are visible to everyone.
        - Personal plans are only visible to the owner.
        - Admins can see all plans.
      operationId: listWorkoutPlans
      parameters:
        - name: is_template
          in: query
          required: false
          description: Filter by template flag (`true` = admin templates only, `false` = personal plans only)
          schema:
            type: boolean
      responses:
        "200":
          description: List of workout plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlan"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Workout Plans]
      summary: Create a new workout plan
      description: |
        Creates a new workout plan with at least one exercise slot.
        - **Admin** can set `is_template = true` to make it a global template.
        - **User** can only create personal plans (`is_template = false`).
      operationId: createWorkoutPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutPlanCreate"
      responses:
        "201":
          description: Workout plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /workout-plans/{planId}:
    get:
      tags: [Workout Plans]
      summary: Get a workout plan by ID
      description: Returns the plan with all its exercise slots. Templates are public; personal plans are owner-only.
      operationId: getWorkoutPlan
      parameters:
        - $ref: "#/components/parameters/PlanIdParam"
      responses:
        "200":
          description: Workout plan with exercise slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Workout Plans]
      summary: Update a workout plan
      description: |
        **Owner or Admin.** Replaces the plan metadata and its full list of exercise slots.
        The request body must include the complete list of `exercises` — any slots not included will be removed.
      operationId: updateWorkoutPlan
      parameters:
        - $ref: "#/components/parameters/PlanIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutPlanCreate"
      responses:
        "200":
          description: Updated workout plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      tags: [Workout Plans]
      summary: Delete a workout plan
      description: "**Owner or Admin.** Permanently deletes the plan and its exercise slots."
      operationId: deleteWorkoutPlan
      parameters:
        - $ref: "#/components/parameters/PlanIdParam"
      responses:
        "204":
          description: Workout plan deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  #  WORKOUT SESSIONS  (Logged gym visits)
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: "HTTP Basic Authentication. Send `Authorization: Basic <base64(username:password)>`."

  # ---------------------------------------------------------------------------
  #  Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    PlanIdParam:
      name: planId
      in: path
      required: true
      description: Unique identifier of the workout plan
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but insufficient permissions for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ---------------------------------------------------------------------------
  #  Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    WorkoutPlan:
      type: object
      description: "A reusable workout template or personal routine (Aggregate Root). Contains one or more exercise slots."
      properties:
        id:
          type: integer
          readOnly: true
          example: 3
        name:
          type: string
          example: Leg Day A
        creator_id:
          type: integer
          readOnly: true
          example: 1
        is_template:
          type: boolean
          description: "`true` = admin-created global template, `false` = personal user plan"
          example: false
        exercises:
          type: array
          description: Ordered list of exercise slots in this plan
          items:
            $ref: "#/components/schemas/PlanExercise"
      required:
        - id
        - name
        - creator_id
        - is_template
        - exercises
    PlanExercise:
      type: object
      description: One exercise slot inside a workout plan (Child Entity)
      properties:
        id:
          type: integer
          readOnly: true
          example: 10
        exercise_id:
          type: integer
          description: References a master Exercise
          example: 42
        order_index:
          type: integer
          description: Position of this exercise in the plan (1-indexed)
          minimum: 1
          example: 1
        target_sets:
          type: integer
          description: Recommended number of sets
          minimum: 1
          example: 4
        target_reps:
          type: integer
          description: Recommended number of reps per set
          minimum: 1
          example: 10
        target_weight:
          type: number
          nullable: true
          description: Optional target weight in **kg**. `null` for bodyweight or unspecified.
          minimum: 0
          example: 80.0
      required:
        - exercise_id
        - order_index
        - target_sets
        - target_reps
    WorkoutPlanCreate:
      type: object
      description: |
        Payload for creating or updating a workout plan (must include at least one exercise slot).
        **Only strength-training exercises are allowed in plans.** Cardio exercises must be logged directly via workout sessions.
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Leg Day A
        is_template:
          type: boolean
          description: "Only admins can set this to `true`"
          default: false
        exercises:
          type: array
          description: Must contain at least one exercise slot
          minItems: 1
          items:
            $ref: "#/components/schemas/PlanExerciseInput"
      required:
        - name
        - exercises
    PlanExerciseInput:
      type: object
      description: Input for one exercise slot (no ID needed — server assigns it)
      properties:
        exercise_id:
          type: integer
          example: 42
        order_index:
          type: integer
          minimum: 1
          example: 1
        target_sets:
          type: integer
          minimum: 1
          example: 4
        target_reps:
          type: integer
          minimum: 1
          example: 10
        target_weight:
          type: number
          nullable: true
          description: Optional target weight in **kg**
          minimum: 0
          example: 80.0
      required:
        - exercise_id
        - order_index
        - target_sets
        - target_reps

    # ── Workout Session Aggregate ──────────────────────────────────────────
...
    Error:
      type: object
      description: Standard error response
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          description: Human-readable error description
          example: The requested resource was not found
      required:
        - error
        - message
    ValidationErrorResponse:
      type: object
      description: Detailed validation error response
      properties:
        error:
          type: string
          example: validation_error
        message:
          type: string
          example: Request body failed validation
        details:
          type: array
          description: List of individual field validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: name
              message:
                type: string
                example: "Field is required"
      required:
        - error
        - message
        - details